requirements.txt
psycopg2~=2.9.6
openai~=0.27.5
python-dotenv~=1.0.0